requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
geoip2==2.9.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
maxminddb==1.4.1
mccabe==0.6.1
privex-helpers==1.1.4
privex-loghelper==1.0.5
pylint==2.3.1
python-dotenv==0.10.3
redis==3.3.6
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
